피드백 반영 - 정규식 수정 - compareTo 수정 - 클래스, 테스트 분리
